Make fast-jar packaging a bit more flexible
Ideally, we should extract the mutable jar from AbstractFastJarBuilder
but it looks like a task for another day.
